Modified SSCrypto(GE) category substantially. Some "old" code remains.
"Old" code has been marked as possibly deprecated and will be removed
in the near future should I decide it is not destined for this world.
Cleaned up external API to be more task-focussed and all-encompassing,
rather than having half a dozen or more smaller, utility functions.

A certificate can now be created by using another certificate as a
"base" (or template, if you will) upon which modifications can be made.

Keys created using this mechanism are not first-class citizens as
decided by the official Security.framework. Keys appear to not be
useful for SSL authentication unless they are explicitly added to
the keychain -- a manual process at this stage.