Created new branch to test out the idea of using OS X's keychain.
I can't quite remember why I avoided this in the first place, so
we'll see where this goes. Keys now play nicely with the rest of
the Security.framework API, so all is well on that front.

Right now the keychain reference is a global variable of file-
scope which does *not* get cleaned up when the application is
closed. Likewise, it is not removed from (unsightly) view in
OS X's built-in Keychain Access application.